feat: Refactor vesting contracts into modular components
- Replaced VestingVault.sol with dedicated vesting token contracts:
  - BaseVestingToken.sol
  - EmployeeVestingToken.sol
  - FounderVestingToken.sol
  - InvestorVestingToken.sol
- Updated test suite to align with the new contract structure.

This refactoring improves modularity and maintainability of the vesting logic.
