+++ test/InvestorVestingToken.test.ts
@@ -1,0 +1,153 @@
+import { ethers } from "hardhat";
+import { expect } from "chai";
+import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
+import { AimondToken, InvestorVestingToken } from "../typechain-types";
+import * as helpers from "@nomicfoundation/hardhat-network-helpers";
+
+import { formatTimestamp, formatAmdBalance, formatAimBalance } from "./utils/time";
+
+describe("InvestorVestingToken Scenarios", function () {
+    async function deployVestingFixture() {
+        const [owner, beneficiary] = await ethers.getSigners();
+        const amdToken = await ethers.deployContract("AimondToken", [owner.address]);
+        const vestingToken = await ethers.deployContract("InvestorVestingToken", [owner.address, await amdToken.getAddress()]);
+        
+        const amdDecimals = await amdToken.decimals();
+
+        // Transfer VestingToken to beneficiary
+        const scheduleAmount = ethers.parseUnits("10001", 18);
+        console.log("Schedule Amount:", formatAimBalance(scheduleAmount));
+        await vestingToken.connect(owner).transfer(beneficiary.address, scheduleAmount);
+
+        // Transfer AMD to vesting contract
+        const totalAmdForVesting = ethers.parseUnits("100000", 18); // Example total amount for vesting
+        await amdToken.connect(owner).transfer(await vestingToken.getAddress(), totalAmdForVesting);
+
+        return { vestingToken, amdToken, owner, beneficiary, scheduleAmount, amdDecimals };
+    }
+
+    it("Should not allow claiming before cliff ends", async function () {
+        const { vestingToken, amdToken, owner, beneficiary, scheduleAmount } = await helpers.loadFixture(deployVestingFixture);
+
+        const listingTimestamp = await helpers.time.latest();
+        console.log("Investor Listing Timestamp:", formatTimestamp(Number(listingTimestamp)), `(${listingTimestamp})`);
+        await vestingToken.connect(owner).createVesting(beneficiary.address);
+        await vestingToken.connect(owner).setGlobalStartTime(listingTimestamp);
+        
+        const schedule = await vestingToken.vestingSchedules(beneficiary.address);
+        const globalStartTime = await vestingToken.globalStartTime();
+        console.log("Global Start Time:", formatTimestamp(Number(globalStartTime)), `(${globalStartTime})`);
+        console.log("Schedule Cliff Duration:", schedule.cliffDuration.toString());
+        const cliffEndsTimestamp = Number(globalStartTime) + Number(schedule.cliffDuration);
+        const twoSecBeforeCliffEnds = cliffEndsTimestamp - 2;
+        console.log("2 Sec Before Cliff Ends Timestamp:", formatTimestamp(Number(twoSecBeforeCliffEnds)), `(${twoSecBeforeCliffEnds})`);
+
+        const initialBalance = await amdToken.balanceOf(beneficiary.address);
+        console.log("Investor AMD Initaial Balance:", formatAmdBalance(initialBalance));
+
+        await helpers.time.increaseTo(twoSecBeforeCliffEnds);
+        console.log("Current Block Timestamp (before claim):", formatTimestamp(Number(await helpers.time.latest())), `(${await helpers.time.latest()})`);
+        
+        const releasableAmount = await vestingToken.getCurrentlyReleasableAmount(beneficiary.address);
+        expect(releasableAmount).to.equal(0);
+
+        await vestingToken.connect(beneficiary).claim();
+        console.log("Investor AMD Balance (Before Cliff Claim):", formatAmdBalance(await amdToken.balanceOf(beneficiary.address)));
+        expect(await amdToken.balanceOf(beneficiary.address)).to.equal(initialBalance);
+    });
+
+    it("Should allow claiming right after cliff ends", async function () {
+        const { vestingToken, amdToken, owner, beneficiary, scheduleAmount } = await helpers.loadFixture(deployVestingFixture);
+
+        const listingTimestamp = await helpers.time.latest();
+        console.log("Investor Listing Timestamp:", formatTimestamp(Number(listingTimestamp)), `(${listingTimestamp})`);
+        await vestingToken.connect(owner).createVesting(beneficiary.address);
+        await vestingToken.connect(owner).setGlobalStartTime(listingTimestamp);
+        
+        const schedule = await vestingToken.vestingSchedules(beneficiary.address);
+        const globalStartTime = await vestingToken.globalStartTime();
+        console.log("Global Start Time:", formatTimestamp(Number(globalStartTime)), `(${globalStartTime})`);
+        const cliffEndsTimestamp = Number(globalStartTime) + Number(schedule.cliffDuration) - 1;
+        console.log("Cliff Ends Timestamp:", formatTimestamp(Number(cliffEndsTimestamp)), `(${cliffEndsTimestamp})`);
+
+        await helpers.time.increaseTo(cliffEndsTimestamp); // one sec before cliff end
+        console.log("Investor Claim Time (After Cliff):", formatTimestamp(await helpers.time.latest()), `(${await helpers.time.latest()})`);
+        await vestingToken.connect(beneficiary).claim();
+
+        const expectedAmount = scheduleAmount / schedule.installmentCount;
+        console.log("Investor AMD Balance (after cliff claim):", formatAmdBalance(await amdToken.balanceOf(beneficiary.address)));
+        expect(await amdToken.balanceOf(beneficiary.address)).to.equal(expectedAmount);
+    });
+
+    it("Should allow claiming one sec after cliff ends", async function () {
+        const { vestingToken, amdToken, owner, beneficiary, scheduleAmount } = await helpers.loadFixture(deployVestingFixture);
+
+        const listingTimestamp = await helpers.time.latest();
+        console.log("Investor Listing Timestamp:", formatTimestamp(Number(listingTimestamp)), `(${listingTimestamp})`);
+        await vestingToken.connect(owner).createVesting(beneficiary.address);
+        await vestingToken.connect(owner).setGlobalStartTime(listingTimestamp);
+        
+        const schedule = await vestingToken.vestingSchedules(beneficiary.address);
+        const globalStartTime = await vestingToken.globalStartTime();
+        console.log("Global Start Time:", formatTimestamp(Number(globalStartTime)), `(${globalStartTime})`);
+        const cliffEndsTimestamp = Number(globalStartTime) + Number(schedule.cliffDuration);
+        console.log("Cliff Ends Timestamp:", formatTimestamp(Number(cliffEndsTimestamp)), `(${cliffEndsTimestamp})`);
+
+        await helpers.time.increaseTo(cliffEndsTimestamp); // Exactly at cliff end
+        console.log("Investor Claim Time (After Cliff):", formatTimestamp(await helpers.time.latest()), `(${await helpers.time.latest()})`);
+        await vestingToken.connect(beneficiary).claim();
+
+        const expectedAmount = scheduleAmount / schedule.installmentCount;
+        console.log("Investor AMD Balance (after cliff claim):", formatAmdBalance(await amdToken.balanceOf(beneficiary.address)));
+        expect(await amdToken.balanceOf(beneficiary.address)).to.equal(expectedAmount);
+    });
+
+    it("Should allow claiming all tokens after full vesting period", async function () {
+        const { vestingToken, amdToken, owner, beneficiary, scheduleAmount, amdDecimals } = await helpers.loadFixture(deployVestingFixture);
+
+        const listingTimestamp = await helpers.time.latest();
+        await vestingToken.connect(owner).createVesting(beneficiary.address);
+        await vestingToken.connect(owner).setGlobalStartTime(listingTimestamp);
+        
+        const schedule = await vestingToken.vestingSchedules(beneficiary.address);
+        const globalStartTime = await vestingToken.globalStartTime();
+        const fullVestingEndsTimestamp = Number(globalStartTime) + Number(schedule.cliffDuration) + Number(schedule.vestingDuration);
+        
+        console.log("Full Vesting Ends Timestamp:", formatTimestamp(Number(fullVestingEndsTimestamp)), `(${fullVestingEndsTimestamp})`);
+
+        await helpers.time.increaseTo(fullVestingEndsTimestamp);
+        console.log("Investor Claim Time (After Full Vesting):", formatTimestamp(await helpers.time.latest()), `(${await helpers.time.latest()})`);
+        await vestingToken.connect(beneficiary).claim();
+
+        console.log("Investor AMD Balance (after full vesting claim):", formatAmdBalance(await amdToken.balanceOf(beneficiary.address)));
+        expect(await amdToken.balanceOf(beneficiary.address)).to.equal(scheduleAmount);
+    });
+
+    it("Should release tokens correctly over each installment", async function () {
+        const { vestingToken, amdToken, owner, beneficiary, scheduleAmount, amdDecimals } = await helpers.loadFixture(deployVestingFixture);
+
+        await vestingToken.connect(owner).createVesting(beneficiary.address);
+        await vestingToken.connect(owner).setGlobalStartTime(await helpers.time.latest());
+
+        const schedule = await vestingToken.vestingSchedules(beneficiary.address);
+        const globalStartTime = await vestingToken.globalStartTime();
+
+        let totalClaimedAmd = BigInt(0);
+
+        const installmentDuration = schedule.vestingDuration / schedule.installmentCount;
+
+        const expectedAmount = scheduleAmount / schedule.installmentCount;
+
+        for (let i = 0; i < schedule.installmentCount; i++) {
+            const newTimestamp = Number(globalStartTime) + Number(schedule.cliffDuration) + Number(installmentDuration) * i - 1;
+            console.log("New installment:", formatTimestamp(Number(newTimestamp)), `(${newTimestamp})`);
+        
+            await helpers.time.increaseTo(newTimestamp);
+            await vestingToken.connect(beneficiary).claim();
+
+            totalClaimedAmd += expectedAmount;
+            console.log("Investor AMD Balance (after installment):", formatAmdBalance(await amdToken.balanceOf(beneficiary.address)));
+            expect(await amdToken.balanceOf(beneficiary.address)).to.equal(totalClaimedAmd);
+        }
+    });
+});
